Move plugin/metricstore to internal/storage/metricstore
