+++ internal/storage/metricstore/prometheus/metricstore/testdata/service_datapoint_response.json
